workshop_10_11_12_13/first_web_project/urls.py
"""
URL configuration for first_web_project project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/5.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path, include
from debug_toolbar.toolbar import debug_toolbar_urls

from core.views import (
    home_view,
    about_view,
    post_details_view,
    post_create_view,
)

urlpatterns = [
    path("admin/", admin.site.urls),
    path("", home_view, name="home"),
    path("post/add/", post_create_view),
    path("post/<int:pk>/", post_details_view),
    path("about/", about_view),
    path("", include("users.urls")),
    # path("", include("core.urls")),

] + debug_toolbar_urls()
