workshop_10_11_12_13/pyproject.toml
[project]
name = "workshop-10"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crispy-tailwind>=1.0.3",
    "django>=6.0",
    "django-crispy-forms>=2.5",
    "django-debug-toolbar>=6.1.0",
]
